Split out SEXP list functions to a separate header
If the modules will be smaller we can in future
reduce their dependencies and then we could enable
the temporarily disabled tests.
