+++ src/OVAL/probes/probe/worker.c
@@ -37,6 +37,7 @@
 #include "worker.h"
 #include "probe-table.h"
 #include "probe.h"
+#include "_sexp-list.h"
 
 extern bool  OSCAP_GSYM(varref_handling);
 extern void *OSCAP_GSYM(probe_arg);
